+++ packages/i18n/I18nDecorator/tests/I18nDecorator-specs.js
@@ -5,7 +5,7 @@ import {shallow, mount} from 'enzyme';
 
 import ilib from '../../ilib/lib/ilib.js';
 import {updateLocale} from '../../locale';
-import {contextTypes, I18nDecorator} from '../I18nDecorator';
+import {I18nContextDecorator, I18nDecorator} from '../I18nDecorator';
 
 describe('I18nDecorator', () => {
 
@@ -20,27 +20,36 @@ describe('I18nDecorator', () => {
 	});
 
 	it('should add rtl context parameter', function () {
-		const Component = (props, context) => (
-			<div>{'rtl' in context ? 'has rtl context' : 'does not have rtl context'}</div>
+		const Component = (props) => (
+			<div>{'rtl' in props ? 'has rtl prop' : 'does not have rtl prop'}</div>
 		);
-		Component.contextTypes = contextTypes;
 
-		const Wrapped = I18nDecorator(Component);
+		const Wrapped = I18nDecorator(
+			I18nContextDecorator(
+				{rtlProp: 'rtl'},
+				Component
+			)
+		);
 		const subject = mount(<Wrapped />);
 
-		const expected = 'has rtl context';
+		const expected = 'has rtl prop';
 		const actual = subject.text();
 
 		expect(actual).to.equal(expected);
 	});
 
 	it('should add updateLocale context parameter', function () {
-		const Component = (props, context) => (
-			<div>{typeof context.updateLocale}</div>
+		// eslint-disable-next-line enact/prop-types
+		const Component = ({updateLocale: update}) => (
+			<div>{typeof update}</div>
 		);
-		Component.contextTypes = contextTypes;
 
-		const Wrapped = I18nDecorator(Component);
+		const Wrapped = I18nDecorator(
+			I18nContextDecorator(
+				{updateLocaleProp: 'updateLocale'},
+				Component
+			)
+		);
 		const subject = mount(<Wrapped />);
 
 		const expected = 'function';
@@ -50,16 +59,21 @@ describe('I18nDecorator', () => {
 	});
 
 	it('should update the current locale when updateLocale is called', function () {
-		const Component = (props, context) => {
-			const handleClick = () => context.updateLocale('ar-SA');
+		// eslint-disable-next-line enact/prop-types
+		const Component = ({_updateLocale}) => {
+			const handleClick = () => _updateLocale('ar-SA');
 
 			return (
 				<button onClick={handleClick} />
 			);
 		};
-		Component.contextTypes = contextTypes;
 
-		const Wrapped = I18nDecorator(Component);
+		const Wrapped = I18nDecorator(
+			I18nContextDecorator(
+				{updateLocaleProp: '_updateLocale'},
+				Component
+			)
+		);
 		const subject = mount(<Wrapped />);
 		subject.find('button').simulate('click');
 
@@ -70,16 +84,21 @@ describe('I18nDecorator', () => {
 	});
 
 	it('should update the rtl context parameter when RTL changes', function () {
-		const Component = (props, context) => {
-			const handleClick = () => context.updateLocale('ar-SA');
+		// eslint-disable-next-line enact/prop-types
+		const Component = ({rtl, _updateLocale}) => {
+			const handleClick = () => _updateLocale('ar-SA');
 
 			return (
-				<button onClick={handleClick}>{context.rtl ? 'rtl' : 'ltr'}</button>
+				<button onClick={handleClick}>{rtl ? 'rtl' : 'ltr'}</button>
 			);
 		};
-		Component.contextTypes = contextTypes;
 
-		const Wrapped = I18nDecorator(Component);
+		const Wrapped = I18nDecorator(
+			I18nContextDecorator(
+				{rtlProp: 'rtl', updateLocaleProp: '_updateLocale'},
+				Component
+			)
+		);
 		const subject = mount(<Wrapped />);
 		const button = subject.find('button');
 
@@ -113,7 +132,7 @@ describe('I18nDecorator', () => {
 
 		const Wrapped = I18nDecorator(Component);
 		// explicitly setting locale so we get a known class list regardless of runtime locale
-		const subject = shallow(<Wrapped locale="en-US" />);
+		const subject = shallow(<Wrapped locale="en-US" />).find(Component);
 
 		const expected = true;
 		const actual =	subject.hasClass('enact-locale-en') &&