+++ packages/moonstone/Input/Input.js
@@ -7,8 +7,7 @@
  */
 
 import kind from '@enact/core/kind';
-import {Subscription} from '@enact/core/internal/PubSub';
-import {contextTypes} from '@enact/i18n/I18nDecorator';
+import {I18nContextDecorator} from '@enact/i18n/I18nDecorator';
 import {isRtlText} from '@enact/i18n/util';
 import Changeable from '@enact/ui/Changeable';
 import Pure from '@enact/ui/internal/Pure';
@@ -226,8 +225,6 @@ const InputBase = kind({
 		type: 'text'
 	},
 
-	contextTypes,
-
 	styles: {
 		css: componentCss,
 		className: 'decorator',
@@ -304,14 +301,15 @@ const InputBase = kind({
  *
  * @class Input
  * @memberof moonstone/Input
+ * @mixes i18n/I18nDecorator.I18nContextDecorator
  * @mixes ui/Changeable.Changeable
  * @mixes spotlight/Spottable.Spottable
  * @ui
  * @public
  */
 const Input = Pure(
-	Subscription(
-		{channels: ['i18n'], mapMessageToProps: (channel, {rtl}) => ({rtl})},
+	I18nContextDecorator(
+		{rtlProp: 'rtl'},
 		Changeable(
 			InputSpotlightDecorator(
 				Skinnable(