+++ packages/moonstone/Spinner/Spinner.js
@@ -18,6 +18,7 @@ import compose from 'ramda/src/compose';
 import React from 'react';
 import Pause from '@enact/spotlight/Pause';
 import UiSpinnerBase from '@enact/ui/Spinner';
+import Spotlight from '@enact/spotlight';
 
 import $L from '../internal/$L';
 import Marquee from '../Marquee';
@@ -176,6 +177,7 @@ const SpinnerSpotlightDecorator = hoc((config, Wrapped) => {
 
 			if (blockClickOn === 'screen') {
 				this.paused.pause();
+				Spotlight.getCurrent().blur();
 			}
 		}
 
@@ -183,6 +185,7 @@ const SpinnerSpotlightDecorator = hoc((config, Wrapped) => {
 			const {blockClickOn} = this.props;
 
 			if (blockClickOn === 'screen') {
+				Spotlight.focus();
 				this.paused.resume();
 			}
 		}