+++ packages/moonstone/RangePicker/RangePicker.js
@@ -102,7 +102,7 @@ const RangePickerBase = kind({
 
 		/**
 		 * When `true`, the RangePicker is shown as disabled and does not generate `onChange`
-		 * [events]{@glossary event}.
+		 * [events]{@link /docs/developer-guide/glossary/#event}.
 		 *
 		 * @type {Boolean}
 		 * @public