+++ packages/ui/CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Added
 
-- `ui/VirtualList` support `data-webos-voice-focused` and `data-webos-voice-group label`
+- `ui/VirtualList` support `data-webos-voice-focused` and `data-webos-voice-group-label`
 
 ## [2.0.0-beta.9] - 2018-07-02