+++ packages/ui/CHANGELOG.md
@@ -2,11 +2,11 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
-## [2.0.0-rc.1]
+## [Unreleased]
 
 ### Added
 
-- `ui/VirtualList` support `data-webos-voice-focused` and `data-webos-voice-group-label`
+- `ui/VirtualList.VirtualList` and `ui/VirtualList.VirtualGridList` support `data-webos-voice-focused`, `data-webos-voice-group-label`
 
 ## [2.0.0-beta.9] - 2018-07-02