+++ packages/ui/CHANGELOG.md
@@ -4,9 +4,16 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ## [Unreleased]
 
+### Removed
+
+- `ui/FloatingLayer.contextTypes` export
+- `ui/Marquee.controlContextTypes` export
+- `ui/Placeholder.contextTypes` export
+- `ui/Resizable.contextTypes` export
+
 ### Added
 
-- `ui/VirtualList.VirtualList` and `ui/VirtualList.VirtualGridList` support `data-webos-voice-focused`, `data-webos-voice-group-label`
+- `ui/VirtualList.VirtualList` and `ui/VirtualList.VirtualGridList` support `data-webos-voice-focused` and `data-webos-voice-group-label`
 
 ## [2.0.0-beta.9] - 2018-07-02